更新组件

+++ src/components/packages/__VUE/numberkeyboard/index.vue
@@ -6,13 +6,13 @@
     :popClass="popClass"
     :overlay="overlay"
     @click-overlay="closeBoard()"
-    :isWrapTeleport="false"
-    overlayClass="nut-numberkeyboard-overlay"
+    :teleportDisable="teleportDisable"
+    overlay-class="nut-numberkeyboard-overlay"
   >
     <div class="nut-numberkeyboard" ref="root">
       <div class="number-board-header" v-if="title">
         <h3 class="tit">{{ title }}</h3>
-        <span class="keyboard-close" @click="closeBoard()">{{ translate('done') }}</span>
+        <span class="keyboard-close" v-if="type == 'default'" @click="closeBoard()">{{ translate('done') }}</span>
       </div>
       <div class="number-board-body">
         <div class="number-board">
@@ -34,16 +34,16 @@
                 { lock: item.type == 'lock' },
                 { delete: item.type == 'delete' }
               ]"
-              @touchstart.stop.prevent="(event) => onTouchstart(item, event)"
-              @touchmove.stop.prevent="(event) => onTouchMove(item, event)"
-              @touchend.stop.prevent="(event) => onTouchEnd(event)"
+              @touchstart="(event) => onTouchstart(item, event)"
+              @touchmove="(event) => onTouchMove(event)"
+              @touchend="(event) => onTouchEnd(event)"
             >
               <template v-if="item.type == 'number' || item.type == 'custom'">{{ item.id }}</template>
-              <image
+              <img
                 v-if="item.type == 'lock'"
                 src="https://img11.360buyimg.com/imagetools/jfs/t1/146371/38/8485/738/5f606425Eca239740/14f4b4f5f20d8a68.png"
               />
-              <image
+              <img
                 v-if="item.type == 'delete'"
                 src="https://img11.360buyimg.com/imagetools/jfs/t1/129395/8/12735/2030/5f61ac37E70cab338/fb477dc11f46056c.png"
               />
@@ -55,15 +55,15 @@
             <div
               :class="['key', { active: clickKeyIndex == 'delete' }]"
               @touchstart="(event) => onTouchstart({ id: 'delete', type: 'delete' }, event)"
-              @touchmove="(event) => onTouchMove({ id: 'delete', type: 'delete' }, event)"
+              @touchmove="(event) => onTouchMove(event)"
               @touchend="onTouchEnd"
             >
-              <image
+              <img
                 src="https://img11.360buyimg.com/imagetools/jfs/t1/129395/8/12735/2030/5f61ac37E70cab338/fb477dc11f46056c.png"
               />
             </div>
           </div>
-          <div class="key-board-wrapper key-board-finish" @click="closeBoard()" v-if="title == ''">
+          <div class="key-board-wrapper key-board-finish" @click="closeBoard()">
             <div :class="['key', 'finish', { activeFinsh: clickKeyIndex == 'finish' }]">
               {{ confirmText || translate('done') }}
             </div>
@@ -76,8 +76,12 @@
 
 <script lang="ts">
 import { computed, onMounted, provide, reactive, nextTick, ref, watch, Ref } from 'vue';
-import { createComponent } from '../../utils/create';
+import { createComponent } from '@/components/packages/utils/create';
 const { create, translate } = createComponent('numberkeyboard');
+export interface keys {
+  id: number | string;
+  type: string;
+}
 export default create({
   props: {
     confirmText: {
@@ -116,12 +120,12 @@ export default create({
       type: Boolean,
       default: true
     },
-    isWrapTeleport: {
+    teleportDisable: {
       type: Boolean,
       default: true
     },
     teleport: {
-      type: [String],
+      type: [String, Element],
       default: 'body'
     },
     popClass: {
@@ -131,9 +135,7 @@ export default create({
   },
   emits: ['input', 'delete', 'close', 'update:value'],
   setup(props, { emit }) {
-    // console.log(props.overlay);
-
-    const clickKeyIndex = ref(undefined);
+    const clickKeyIndex: Ref<string | undefined | number> = ref(undefined);
     const show = ref(props.visible);
     const root = ref<HTMLElement>();
     function defaultKey() {
@@ -153,7 +155,7 @@ export default create({
     }
 
     function getBasicKeys() {
-      const keys: any = [];
+      const keys: keys[] = [];
       for (let i = 1; i <= 9; i++) {
         keys.push({ id: i, type: 'number' });
       }
@@ -171,8 +173,11 @@ export default create({
       if (customKeys.length > 2) {
         customKeys = [customKeys[0], customKeys[1]];
       }
+      if (customKeys.length == 2 && props.title && props.type != 'rightColumn') {
+        customKeys = [customKeys[0]];
+      }
       if (customKeys.length === 1) {
-        if (props.title) {
+        if (props.title && props.type != 'rightColumn') {
           keys.push({ id: customKeys[0], type: 'custom' }, { id: 0, type: 'number' }, { id: 'delete', type: 'delete' });
         } else {
           keys.push({ id: 0, type: 'number' }, { id: customKeys[0], type: 'custom' });
@@ -183,11 +188,6 @@ export default create({
           { id: 0, type: 'number' },
           { id: customKeys[1], type: 'custom' }
         );
-        if (props.title) {
-          keys.push({ id: 'delete', type: 'delete' });
-        }
-      } else {
-        keys.push({ id: 0, type: 'number' });
       }
       return keys;
     }
@@ -204,7 +204,7 @@ export default create({
       }
     );
 
-    function onTouchstart(item: any, event: any) {
+    function onTouchstart(item: { id: string | number; type: string }, event: TouchEvent) {
       event.stopPropagation();
       clickKeyIndex.value = item.id;
       if (item.type == 'number' || item.type == 'custom') {
@@ -221,10 +221,10 @@ export default create({
         emit('update:value', props.value.slice(0, props.value.length - 1));
       }
     }
-    function onTouchMove(id: any, event: any) {
+    function onTouchMove(event: TouchEvent) {
       event.stopPropagation();
     }
-    function onTouchEnd(event: any) {
+    function onTouchEnd(event: TouchEvent) {
       event.preventDefault();
       clickKeyIndex.value = undefined;
     }
@@ -233,7 +233,6 @@ export default create({
       emit('close');
     }
 
-    onMounted(() => {});
     return {
       clickKeyIndex,
       defaultKey,
@@ -252,5 +251,5 @@ export default create({
 });
 </script>
 <style lang="scss">
-@import './index.scss'
+@import './index.scss' 
 </style>
